Implement Cuenta and Movimiento services with mappers and controllers

+++ src/main/java/com/ntt/challenge/dto/MovimientoRequestDTO.java
@@ -2,11 +2,12 @@ package com.ntt.challenge.dto;
 
 import com.ntt.challenge.model.TipoMovimiento;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record MovimientoRequestDTO(
         UUID cuentaId,
         TipoMovimiento tipoMovimiento,
-        Double valor
+        BigDecimal valor
 ) {
 }
